fix(test): cast INT16 extremes so negate specialization is used
Bare INT16_MAX/MIN are int-ranked macros, so negate() took the generic
template and failed the extremal checks in CI.

+++ src/lib/mathlib/math/FunctionsTest.cpp
@@ -260,7 +260,7 @@ TEST(FunctionsTest, negate)
 
 	EXPECT_EQ(negate(static_cast<int16_t>(5)), static_cast<int16_t>(-5));
 	EXPECT_EQ(negate(static_cast<int16_t>(-5)), static_cast<int16_t>(5));
-	// int16_t extremal values swap as documented
-	EXPECT_EQ(negate(INT16_MAX), INT16_MIN);
-	EXPECT_EQ(negate(INT16_MIN), INT16_MAX);
+	// Cast required: INT16_* macros are int-ranked; bare value skips int16 specialization
+	EXPECT_EQ(negate(static_cast<int16_t>(INT16_MAX)), static_cast<int16_t>(INT16_MIN));
+	EXPECT_EQ(negate(static_cast<int16_t>(INT16_MIN)), static_cast<int16_t>(INT16_MAX));
 }
